feat(girth): adding nested stack and lambdas access to fargate
Refs: #1667
