+++ packages/infra/lib/api-stack.ts
@@ -51,6 +51,7 @@ import { getSecrets, Secrets } from "./shared/secrets";
 import { provideAccessToQueue } from "./shared/sqs";
 import { isProd, isSandbox, mbToBytes } from "./shared/util";
 import { wafRules } from "./shared/waf-rules";
+import { GirthLambdasNestedStack } from "./girth-stack";
 
 const FITBIT_LAMBDA_TIMEOUT = Duration.seconds(60);
 const CDA_TO_VIS_TIMEOUT = Duration.minutes(15);
@@ -448,6 +449,20 @@ export class APIStack extends Stack {
         configId: appConfigConfigId,
       },
       cookieStore,
+    });
+    new GirthLambdasNestedStack(this, "GirthLambdasNestedStack", {
+      lambdaLayers,
+      vpc: this.vpc,
+      apiService: apiService,
+      secrets,
+      cqOrgCertificate: props.config.carequality.secretNames.CQ_ORG_CERTIFICATE,
+      cqOrgPrivateKey: props.config.carequality.secretNames.CQ_ORG_PRIVATE_KEY,
+      cqOrgCertificateIntermediate:
+        props.config.carequality.secretNames.CQ_ORG_CERTIFICATE_INTERMEDIATE,
+      cqOrgPrivateKeyPassword: props.config.carequality.secretNames.CQ_ORG_PRIVATE_KEY_PASSWORD,
+      apiURL: apiService.loadBalancer.loadBalancerDnsName,
+      envType: props.config.environmentType,
+      sentryDsn: props.config.lambdasSentryDSN,
     });
 
     // Access grant for Aurora DB
